code clean up in accounts app

+++ accounts/views.py
@@ -1,22 +1,40 @@
-from django.shortcuts import render, redirect
-from django.contrib.auth.models import User
-from django.contrib.auth import authenticate, logout as auth_logout, login as auth_login
-from django.contrib.auth.decorators import login_required
-from django.contrib.sites.shortcuts import get_current_site
-from django.template.loader import render_to_string
-from django.utils.http import urlsafe_base64_encode, urlsafe_base64_decode
-from django.utils.encoding import force_bytes, force_text, DjangoUnicodeDecodeError
-from .utils import generate_token
-from .models import StudentProfile, TutorProfile, Countries, SocialConnection, UserSession
-from tutoring.models import QuestionAnswer
-from django.core.mail import EmailMessage
+# from accounts.seed_data_installer import *
+from accounts.models import Countries
+from accounts.models import SocialConnection
+from accounts.models import StudentProfile
+from accounts.models import TutorProfile
+from accounts.models import UserSession
+from accounts.utils import generate_token
 from django.conf import settings
 from django.contrib import messages
+from django.contrib.auth import authenticate
+from django.contrib.auth import login as auth_login
+from django.contrib.auth import logout as auth_logout
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.models import User
+from django.contrib.sites.shortcuts import get_current_site
 from django.core.cache import cache
+from django.core.mail import EmailMessage
 from django.http import HttpResponse
-import json, re, os, random, string, requests, datetime
-from .seed_data_installer import *
+from django.shortcuts import redirect
+from django.shortcuts import render
+from django.template.loader import render_to_string
+from django.utils.encoding import DjangoUnicodeDecodeError
+from django.utils.encoding import force_bytes
+from django.utils.encoding import force_text
+from django.utils.http import urlsafe_base64_decode
+from django.utils.http import urlsafe_base64_encode
 from http import HTTPStatus
+from tutoring.models import QuestionAnswer
+import accounts.seed_data_installer
+import datetime
+import json
+import onetutor.com.accounts.ValueSet as ValueSet
+import os
+import random
+import re
+import requests
+import string
 
 def login(request):
 	if request.method == "POST":
@@ -24,7 +42,7 @@ def login(request):
 		if cache.get('loginAttempts') != None and cache.get('loginAttempts') > 3:
 			cache.set('loginAttempts', cache.get('loginAttempts'), 600)
 			context = {"message": 'Your account has been temporarily locked out because of too many failed login attempts.'}
-			return render(request,"accounts/login.html", context)
+			return render(request,ValueSet.ACCOUNTS_LOGIN_TEMPLATE, context)
 
 		username = request.POST['username'].replace(" ", "")
 		password = request.POST['password']
@@ -48,8 +66,8 @@ def login(request):
 				cache.incr('loginAttempts', 1)
 
 			context = {"message": "Username or Password did not match!", "username": username}
-			return render(request,"accounts/login.html", context)
-	return render(request,"accounts/login.html", {})
+			return render(request,ValueSet.ACCOUNTS_LOGIN_TEMPLATE, context)
+	return render(request,ValueSet.ACCOUNTS_LOGIN_TEMPLATE, {})
 
 def authenticate_user(request, user):
 	response = requests.get("http://ip-api.com/json").json()
@@ -65,7 +83,7 @@ def authenticate_user(request, user):
 		"message": "This IP has been blocked by OneTutor for some reasons. If you think there has been some mistake, please appeal.",
 		"username": user.username
 	}
-	return render(request,"accounts/login.html", context)
+	return render(request,ValueSet.ACCOUNTS_LOGIN_TEMPLATE, context)
 
 def add_user_session(request, browser_type):
 	response = requests.get("http://ip-api.com/json").json()
@@ -88,7 +106,6 @@ def add_user_session(request, browser_type):
 			location=location,
 			ip_address=response['query'],
 		)
-	return
 
 def register(request):
 	if request.method == "POST":
@@ -105,7 +122,7 @@ def register(request):
 				"firstname": firstname,
 				"lastname": lastname
 			}
-			return render(request,"accounts/registration.html", context)
+			return render(request,ValueSet.ACCOUNTS_REGISTRATION_TEMPLATE, context)
 		else:
 			context = {}
 			if(password!=password_2):
@@ -115,7 +132,7 @@ def register(request):
 					"firstname": firstname,
 					"lastname": lastname
 				}
-				return render(request,"accounts/registration.html", context)
+				return render(request,ValueSet.ACCOUNTS_REGISTRATION_TEMPLATE, context)
 
 			if(len(password)<8 or any(letter.isalpha() for letter in password)==False or any(capital.isupper() for capital in password)==False or any(number.isdigit() for number in password)==False):
 				context = {
@@ -124,7 +141,7 @@ def register(request):
 					"firstname": firstname,
 					"lastname": lastname
 				}
-				return render(request,"accounts/registration.html", context)
+				return render(request,ValueSet.ACCOUNTS_REGISTRATION_TEMPLATE, context)
 
 			user = User.objects.create_user(username=email, email=email, password=password, first_name=firstname, last_name=lastname)
 			user.is_active = settings.DEBUG
@@ -145,8 +162,8 @@ def register(request):
 				email_message.send()
 
 			context = {"activate": "We've sent you an activation link. Please check your email."}
-			return render(request,"accounts/registration.html", context)
-	return render(request,"accounts/registration.html", {})
+			return render(request,ValueSet.ACCOUNTS_REGISTRATION_TEMPLATE, context)
+	return render(request,ValueSet.ACCOUNTS_REGISTRATION_TEMPLATE, {})
 
 def logout(request):
 	auth_logout(request)
@@ -163,10 +180,9 @@ def createprofile(request):
 		summary = request.POST["summary"]
 		about = request.POST["about"]
 		subjects = request.POST["subjects"]
-		numberOfEducation = request.POST["numberOfEducation"]
 
 		education = {}
-		for i in range(int(numberOfEducation)):
+		for i in range(int( request.POST["numberOfEducation"])):
 			education["education_" + str(i + 1) ] = {"school_name": request.POST["school_name_" + str(i + 1)], "qualification": request.POST["qualification_" + str(i + 1)], "year": request.POST["year_" + str(i + 1)]}
 
 		availability = {}
@@ -192,7 +208,7 @@ def createprofile(request):
 		}
 
 		TutorProfile.objects.create(user=request.user, summary=summary, about=about, location=location, education=education, subjects=subjects, availability=availability)
-		return redirect("accounts:createprofile")
+		return redirect(ValueSet.ACCOUNTS_CREATE_PROFILE_REDIRECT)
 
 	if request.method == "POST" and "student" in request.POST:
 		# user is a student
@@ -205,7 +221,7 @@ def tutorprofile(request):
 	try:
 		tutorProfile = TutorProfile.objects.get(user=request.user.id)
 	except TutorProfile.DoesNotExist:
-		return redirect("accounts:createprofile")
+		return redirect(ValueSet.ACCOUNTS_CREATE_PROFILE_REDIRECT)
 	
 	tutorProfile.subjects = tutorProfile.subjects.replace(", ", ",").split(",")
 	return render(request,"accounts/tutorprofile.html", {"tutorProfile": tutorProfile})
@@ -215,7 +231,7 @@ def user_settings(request):
 	try:
 		tutorProfile = TutorProfile.objects.get(user=request.user.id)
 	except TutorProfile.DoesNotExist:
-		return redirect("accounts:createprofile")
+		return redirect(ValueSet.ACCOUNTS_CREATE_PROFILE_REDIRECT)
 
 	countries = Countries.objects.all()
 	social_links = SocialConnection.objects.get(user=request.user) if SocialConnection.objects.filter(user=request.user).count() != 0 else None
@@ -241,7 +257,7 @@ def user_settings(request):
 		user.save()
 
 		messages.add_message(request,messages.SUCCESS,"Your personal details has been updated successfully")
-		return redirect("accounts:user_settings")
+		return redirect(ValueSet.ACCOUNTS_USER_SETTINGS_REDIRECT_VIEW)
 
 	if request.method == "POST" and "update_address" in request.POST:
 		address_1 = request.POST["address_1"].strip().title()
@@ -262,17 +278,16 @@ def user_settings(request):
 		}
 		TutorProfile.objects.filter(user=request.user.id).update(location=location)
 		messages.add_message(request,messages.SUCCESS,"Your location has been updated successfully")
-		return redirect("accounts:user_settings")
+		return redirect(ValueSet.ACCOUNTS_USER_SETTINGS_REDIRECT_VIEW)
 
 	if request.method == "POST" and "delete_account" in request.POST:
 		delete_code = request.POST["delete-code"]
-		if "user_" + str(request.user.id) + "_delete_key" in request.session:
-			if delete_code == request.session["user_" + str(request.user.id) + "_delete_key"]:
-				u = User.objects.get(pk=request.user.id)
-				u.delete()
-				del request.session["user_" + str(request.user.id) + "_delete_key"]
-				messages.add_message(request,messages.SUCCESS,"Account deleted successfully")
-				return redirect('tutoring:mainpage')
+		if "user_" + str(request.user.id) + "_delete_key" in request.session and delete_code == request.session["user_" + str(request.user.id) + "_delete_key"]:
+			u = User.objects.get(pk=request.user.id)
+			u.delete()
+			del request.session["user_" + str(request.user.id) + "_delete_key"]
+			messages.add_message(request,messages.SUCCESS,"Account deleted successfully")
+			return redirect('tutoring:mainpage')
 
 	if request.method == "POST" and "update_password" in request.POST:
 		currentPassword = request.POST["currentPassword"]
@@ -283,16 +298,16 @@ def user_settings(request):
 
 		if currentPassword and not user.check_password(currentPassword):
 			messages.add_message(request,messages.ERROR,"Your current password does not match")
-			return redirect("accounts:user_settings")
+			return redirect(ValueSet.ACCOUNTS_USER_SETTINGS_REDIRECT_VIEW)
 
 		if(newPassword and confirmPassword):
 			if(newPassword!=confirmPassword):
 				messages.add_message(request,messages.ERROR,"Your new password and confirm password does not match")
-				return redirect("accounts:user_settings")
+				return redirect(ValueSet.ACCOUNTS_USER_SETTINGS_REDIRECT_VIEW)
 
 			if(len(newPassword)<8 or any(letter.isalpha() for letter in newPassword)==False or any(capital.isupper() for capital in newPassword)==False or any(number.isdigit() for number in newPassword)==False):
 				messages.add_message(request,messages.WARNING,"Your new password is not strong enough")
-				return redirect("accounts:user_settings")
+				return redirect(ValueSet.ACCOUNTS_USER_SETTINGS_REDIRECT_VIEW)
 
 			user.set_password(newPassword)
 		user.save()
@@ -312,7 +327,6 @@ def user_settings(request):
 		answer_on_forum = True if request.POST.get('answer_on_forum') else False
 		message_on_chat = True if request.POST.get('message_on_chat') else False
 		# Need to create a Notifications table for each user and add it to event listner.
-		print("true")
 
 	if request.method == "POST" and "social_links" in request.POST:
 		twitter = request.POST["twitter"]
@@ -324,7 +338,7 @@ def user_settings(request):
 			defaults={'twitter': twitter, 'facebook': facebook, 'google': google, 'linkedin': linkedin},
 		)
 		messages.add_message(request,messages.SUCCESS,"Your social connection has been updated successfully")
-		return redirect("accounts:user_settings")
+		return redirect(ValueSet.ACCOUNTS_USER_SETTINGS_REDIRECT_VIEW)
 
 	if request.is_ajax():
 		functionality = request.GET.get('functionality', None)
@@ -369,7 +383,6 @@ def tutorprofileedit(request):
 		subjects.sort()
 		subjects = ", ".join(subjects)
 		
-		numberOfEducation = request.POST["numberOfEducation"]
 		availabilityChoices = request.POST.getlist('availabilityChoices')
 
 		availability = {}
@@ -388,7 +401,7 @@ def tutorprofileedit(request):
 			availability[weekDay][dayTime] = True
 
 		education = {}
-		for i in range(int(numberOfEducation)):
+		for i in range(int(request.POST["numberOfEducation"])):
 			education["education_" + str(i + 1) ] = {"school_name": request.POST["school_name_" + str(i + 1)], "qualification": request.POST["qualification_" + str(i + 1)], "year": request.POST["year_" + str(i + 1)]}
 
 		TutorProfile.objects.filter(user=request.user.id).update(user=request.user, summary=summary, about=about, education=education, subjects=subjects, availability=availability, profilePicture=None)
@@ -415,8 +428,6 @@ def profile(request):
 
 	if StudentProfile.objects.filter(user=request.user.id).exists():
 		return render("accounts:studentprofile")
-
-	return
 
 def rules(request, rule_type):
 	if rule_type == "privacy_policy":
@@ -485,18 +496,18 @@ def password_change(request, uidb64, token):
 		if password_1 and password_2:
 			if(password_1!=password_2):
 				context = {"message": "Your passwords do not match!"}
-				return render(request,"accounts/password_reset_form.html", context)
+				return render(request,ValueSet.ACCOUNTS_PASSWORD_RESET_FORM_TEMPLATE, context)
 
 			if(len(password_1)<8 or any(letter.isalpha() for letter in password_1)==False or any(capital.isupper() for capital in password_1)==False or any(number.isdigit() for number in password_1)==False):
 				context = {"message": "Your password is not strong enough."}
-				return render(request,"accounts/password_reset_form.html", context)
+				return render(request,ValueSet.ACCOUNTS_PASSWORD_RESET_FORM_TEMPLATE, context)
 
 			user.set_password(password_1)
 			user.save()
 			return redirect("accounts:login")
 
 	if user is not None and generate_token.check_token(user, token):
-		return render(request, "accounts/password_reset_form.html",{})
+		return render(request, ValueSet.ACCOUNTS_PASSWORD_RESET_FORM_TEMPLATE,{})
 	else:
 		return render(request, "accounts/activate_failed.html",status=401)
